I changed my mind; it's better if the main pages import from the [lang] pages.. easier to make sure they exist that way.
